emotion-pipeline/requirements.txt
daily-python>=0.13.0
websockets>=12.0
